refactor(settings): break kube panel to components
