+++ app/react/portainer/settings/SettingsView/KubeSettingsPanel/types.ts
@@ -1,0 +1,12 @@
+export interface FormValues {
+  helmRepositoryURL: string;
+  kubeconfigExpiry: string;
+  globalDeploymentOptions: {
+    hideAddWithForm: boolean;
+    perEnvOverride: boolean;
+    hideWebEditor: boolean;
+    hideFileUpload: boolean;
+    requireNoteOnApplications: boolean;
+    minApplicationNoteLength: number;
+  };
+}
