+++ app/react/portainer/settings/SettingsView/useToggledValue.tsx
